Move vendor registry
